Rename home-revision components dir to home
