+++ context/app/static/js/components/home/AssayTypeBarChart/AssayTypeBarChart.jsx
