+++ context/app/static/js/components/home/AssayTypeBarChart/index.js
