+++ context/app/static/js/components/home/AssayTypeBarChartContainer/AssayTypeBarChartContainer.jsx
