+++ context/app/static/js/components/home/AssayTypeBarChartContainer/index.js
