+++ context/app/static/js/components/home/AssayTypeBarChartContainer/style.js
