+++ context/app/static/js/components/home/AssayTypeBarChartContainer/utils.js
