+++ context/app/static/js/components/home/AssayTypeBarChartDropdown/index.js
