+++ context/app/static/js/components/home/AssayTypeBarChartDropdown/style.js
