+++ context/app/static/js/components/home/EntityCount/index.js
