+++ context/app/static/js/components/home/FacetSearchEntityListItems/style.js
