+++ context/app/static/js/components/home/ImageCarousel/ImageCarousel.jsx
