+++ context/app/static/js/components/home/ImageCarouselContainer/index.js
