+++ context/app/static/js/components/home/ImageCarouselContainer/style.js
