+++ context/app/static/js/components/home/ImageCarouselControlButtons/index.js
