+++ context/app/static/js/components/home/ImageCarouselControlButtons/style.js
