+++ context/app/static/js/pages/Home/Home.jsx
@@ -3,15 +3,15 @@ import { useTheme } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import Typography from '@material-ui/core/Typography';
 
-import AssayTypeBarChartContainer from 'js/components/home-revision/AssayTypeBarChartContainer';
-import ImageCarouselContainer from 'js/components/home-revision/ImageCarouselContainer';
-import Title from 'js/components/home-revision/Title';
-import HuBMAPDescription from 'js/components/home-revision/HuBMAPDescription';
-import EntityCounts from 'js/components/home-revision/EntityCounts';
-import DataUseGuidelines from 'js/components/home-revision/DataUseGuidelines';
-import TwitterTimeline from 'js/components/home-revision/TwitterTimeline';
-import ExternalLinks from 'js/components/home-revision/ExternalLinks';
-import FacetSearch from 'js/components/home-revision/FacetSearch';
+import AssayTypeBarChartContainer from 'js/components/home/AssayTypeBarChartContainer';
+import ImageCarouselContainer from 'js/components/home/ImageCarouselContainer';
+import Title from 'js/components/home/Title';
+import HuBMAPDescription from 'js/components/home/HuBMAPDescription';
+import EntityCounts from 'js/components/home/EntityCounts';
+import DataUseGuidelines from 'js/components/home/DataUseGuidelines';
+import TwitterTimeline from 'js/components/home/TwitterTimeline';
+import ExternalLinks from 'js/components/home/ExternalLinks';
+import FacetSearch from 'js/components/home/FacetSearch';
 import OutboundLink from 'js/shared-styles/Links/OutboundLink';
 
 import {
